fix: divide types file
